Adding rt patches provided by OSADL
Refer to patches.sh to see how the patches were acquired and applied.
This script comes from the "Generate download and patch script" button
at the bottom of OSADL's server webpage: https://www.osadl.org/?id=4075
That machine is in their test farm that is close to the E2E Test
Server's configuration. The patches also enable rt-tests to be run like
cyclictest.